pyproject.toml
[tool.poetry]
name = "ds-bootcamp2025"
version = "0.1.0"
description = ""
authors = ["RandiSPB <swat2299@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standart"], version = "^0.115.12"}
uvicorn = {extras = ["standart"], version = "^0.34.1"}
openai = "^1.75.0"
python-dotenv = "^1.1.0"
openai-agents = "^0.0.11"
pyyaml = "^6.0.2"
jinja2 = "^3.1.6"
websockets = "^15.0.1"
asyncpg = "^0.30.0"
pydantic-settings = "^2.9.0"
python-jose = "^3.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
